requirements.txt
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
lxml==4.9.3
pydantic==1.10.12
python-dotenv==1.0.0
python-telegram-bot==20.4
requests==2.31.0
schedule==1.2.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
typing_extensions==4.7.1
urllib3==2.0.4
